% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myConstructor.R
\name{myConstructor}
\alias{myConstructor}
\title{myConstructor}
\usage{
myConstructor(x, y, alpha)
}
\arguments{
\item{x}{a vector of values}

\item{y}{a vector of values}

\item{alpha}{a significance level less than 1}
}
\value{
returns a list of the x & y vectors in a data frame, the alpha value, the p-value of the test test, and the confidence interval for the difference of the means
}
\description{
myConstructor
}
\examples{
\dontrun{myConstructor(x, y, 0.05)}
}
