% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRttest.R
\name{plot.Rttest}
\alias{plot.Rttest}
\title{plot.Rttest}
\usage{
\method{plot}{Rttest}(x, ...)
}
\arguments{
\item{x}{an Rttest object}

\item{...}{anything else}
}
\value{
prints out a side-by-side boxplot of the variables in the Rttest dataframe
}
\description{
plot.Rttest
}
\examples{
\dontrun{plot(x, ...)}
}
