% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printRttest.R
\name{print.Rttest}
\alias{print.Rttest}
\title{print.Rttest}
\usage{
\method{print}{Rttest}(x, ...)
}
\arguments{
\item{x}{an object of class Rttest}

\item{...}{anything else}
}
\value{
returns the data frame as a kable table, as well as various parameters (p-value, alpha, confidenc interval)
}
\description{
print.Rttest
}
\examples{
\dontrun{print.Rttest(x)}
}
